backend/requirements.txt
Django==3.2
gunicorn==20.1
psycopg2-binary==2.9
djoser==2.2
django-cors-headers==3.11.0
django-filter==23.2
